feat: error context
