+++ gui/src/main/app-layout/error-message.tsx
@@ -1,13 +1,15 @@
-import type { LocalizationKey } from "@/main/i18n-config";
-import { AuthMachineContext } from "@/machines/auth-machine";
-import { TrackingMachineContext } from "@/machines/tracking-machine";
-import { errorsx } from "@@/go/models";
-import { Icon } from "@iconify/react";
-import { useSelector } from "@xstate/react";
-import clsx from "clsx";
-import { useAnimate } from "framer-motion";
 import React from "react";
+import clsx from "clsx";
+import { Icon } from "@iconify/react";
+import { useAnimate } from "framer-motion";
 import { useTranslation } from "react-i18next";
+
+import type { LocalizationKey } from "@/main/i18n-config";
+import type { errorsx } from "@@/go/models";
+
+type ErrorContextType = [error: errorsx.FrontEndError | null, setError: (error: errorsx.FrontEndError | null) => void]
+const ErrorContext = React.createContext<ErrorContextType | null>(null)
+export const useErrorMessage = () => React.useContext(ErrorContext)[1]
 
 const LocalizedErrorMessage: Record<number, LocalizationKey> = {
   401: "errUnauthorized",
@@ -15,16 +17,9 @@ const LocalizedErrorMessage: Record<number, LocalizationKey> = {
   500: "errInternalServerError",
 };
 
-export const ErrorMessage: React.FC= () => {
+export const ErrorMessageProvider: React.FC<React.PropsWithChildren> = ({ children }) => {
   const { t } = useTranslation();
   const [scope, animate] = useAnimate();
-  
-  const authActor = AuthMachineContext.useActorRef()
-  const trackingActor = TrackingMachineContext.useActorRef()
-
-  const authError = useSelector(authActor, ({ context }) => context.error)
-  const trackingError = useSelector(trackingActor, ({ context }) => context.error)
-
   const [error, setError] = React.useState<errorsx.FrontEndError>(null)
 
   React.useEffect(() => {
@@ -38,38 +33,33 @@ export const ErrorMessage: React.FC= () => {
     });
   }, [error]);
 
-  React.useEffect(() => {
-    authError && setError(authError)
-  }, [authError])
-
-  React.useEffect(() => {
-    trackingError && setError(trackingError)
-  }, [trackingError])
-
   return (
-    <div ref={scope} className="absolute">
-      <div
-        id="error-message"
-        className={clsx(
-          "flex gap-6 items-center justify-around",
-          "fixed bottom-2 z-50",
-          "px-8 py-3 rounded-r-xl text-xl backdrop-blur-sm pointer-events-none",
-          "bg-[rgba(255,0,0,.125)]"
-        )}
-        style={{ opacity: 0 }}
-      >
-        <Icon
-          icon="material-symbols:warning-outline"
-          className="text-[#ff6388] w-8 h-8 animate-blink-pulse"
-        />
-        {error && (
-          <span>
-            {LocalizedErrorMessage[error.code]
-              ? t(LocalizedErrorMessage[error.code])
-              : error.message}
-          </span>
-        )}
+    <ErrorContext.Provider value={[error, setError]}>
+      <div ref={scope} className="absolute">
+        <div
+          id="error-message"
+          className={clsx(
+            "flex gap-6 items-center justify-around",
+            "fixed bottom-2 z-50",
+            "px-8 py-3 rounded-r-xl text-xl backdrop-blur-sm pointer-events-none",
+            "bg-[rgba(255,0,0,.125)]"
+          )}
+          style={{ opacity: 0 }}
+        >
+          <Icon
+            icon="material-symbols:warning-outline"
+            className="text-[#ff6388] w-8 h-8 animate-blink-pulse"
+          />
+          {error && (
+            <span>
+              {LocalizedErrorMessage[error.code]
+                ? t(LocalizedErrorMessage[error.code])
+                : error.message}
+            </span>
+          )}
+        </div>
       </div>
-    </div>
+      {children}
+    </ErrorContext.Provider>
   );
 };
