+++ gui/src/pages/output/output-page.tsx
@@ -3,11 +3,12 @@ import { Trans, useTranslation } from "react-i18next";
 import { motion, useAnimate } from "framer-motion";
 import { Icon } from "@iconify/react";
 
+import Dialog from "@/ui/dialog";
+import { Checkbox } from "@/ui/checkbox";
 import { PageHeader } from "@/ui/page-header";
 import { ActionButton } from "@/ui/action-button";
+import { useErrorMessage } from "@/main/app-layout/error-message";
 import { GetThemeList, OpenResultsDirectory } from "@@/go/core/CommandHandler";
-import { Checkbox } from "@/ui/checkbox";
-import Dialog from "@/ui/dialog";
 
 import defaultTheme from './themes/default.png'
 import bladesTheme from './themes/blades.png'
@@ -41,10 +42,11 @@ export const OutputPage: React.FC = () => {
   const [options, setOptions] = React.useState(defaultOptions);
   const [themes, setThemes] = React.useState<string[]>([]);
   const [scope, animate] = useAnimate();
+  const setError = useErrorMessage();
 
   React.useEffect(() => {
     if (themes.length == 0) {
-      GetThemeList().then((themes) => setThemes(themes));
+      GetThemeList().then((themes) => setThemes(themes)).catch(err => setError(err));
     }
   }, []);
