+++ gui/src/pages/tracking/tracking-page.tsx
@@ -1,5 +1,7 @@
 import React from "react";
+import { useSelector } from "@xstate/react";
 import { useTranslation } from "react-i18next";
+
 import { TrackingMachineContext } from "@/machines/tracking-machine";
 import { AuthMachineContext } from "@/machines/auth-machine";
 import { PageHeader } from "@/ui/page-header";
@@ -7,7 +9,7 @@ import { PageHeader } from "@/ui/page-header";
 import { TrackingGamePicker } from "./tracking-game-picker";
 import { TrackingLiveUpdater } from "./tracking-live-updater";
 import { TrackingForm } from "./tracking-form";
-import { useSelector } from "@xstate/react";
+import { useErrorMessage } from "@/main/app-layout/error-message";
 
 export const TrackingPage: React.FC = () => {
   const { t } = useTranslation();
@@ -17,6 +19,19 @@ export const TrackingPage: React.FC = () => {
 
   const authState = useSelector(authActor, (snapshot) => snapshot.value)
   const trackingState = useSelector(trackingActor, (snapshot) => snapshot.value)
+
+  const authError = useSelector(authActor, ({ context }) => context.error)
+  const trackingError = useSelector(trackingActor, ({ context }) => context.error)
+
+  const setError = useErrorMessage()
+
+  React.useEffect(() => {
+    authError && setError(authError)
+  }, [authError])
+
+  React.useEffect(() => {
+    trackingError && setError(trackingError)
+  }, [trackingError])
 
   switch (authState) {
     case "gameForm":
